requirements.txt
django==3.2
djangorestframework==3.12.2
django-ratelimit==3.0.1
django-extensions==3.1.3
psycopg2==2.8.6
python-slugify==4.0.1
pillow==8.1.0
geopy==2.1.0
pytest-cov==2.11.1
pytest-django==4.2.0
pytest-asyncio==0.15.1
channels==3.0.3
channels_redis==3.2.0
websocket-client==0.58.0
selenium==3.141.0
requests==2.25.1
